updates for the club members & matched Headers
I'm stopping here for the delivery, the only update that can be made is for the club members (very time consuming for something that is worth 5/100 point), but you can add multiple types (familyMember, personnel, secondary family member, clubMembers, locations). You just need xampp (for Apache server) in order to make the updates on your devices. you also need to fix the paths so that it matches the structure.